requirement.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.4
dataclasses>=0.6
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.16
django-extensions==3.2.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==4.8.3
jeepney==0.7.1
Jinja2==3.0.3
jwcrypto==1.4.2
MarkupSafe==2.0.1
py==1.11.0
pycparser==2.21
PyJWT>=2.6.0
pytz==2022.6
requests==2.27.1
requests-toolbelt==0.10.1
resilient==47.1.3780
resilient-lib==47.1.3780
retry2==0.9.4
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.13
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
gunicorn>=20.0.0
psycopg2-binary==2.9.5
